requirements.txt
# ============================================================
# FrameAlbum v0.1 - 依赖清单
# Python 3.8+
# 安装：pip install -r requirements.txt
# ============================================================

# Web 框架
Flask==2.3.0
Flask-Cors==4.0.0

# 桌面窗口
pywebview==6.1

# 图像处理
Pillow==12.1.1
pillow_heif==1.3.0

# 日期解析
python-dateutil==2.8.2

# 数据库 ORM
SQLAlchemy==2.0.30

# Windows 桌面通知（可选，仅 Windows；导入完成后弹出 Toast）
# win10toast>=1.0.3

# ============================================================
# 测试依赖（开发环境）
# 安装：pip install -r requirements-dev.txt
# ============================================================
# pytest>=7.0.0
# pytest-cov>=4.0.0
# pytest-mock>=3.0.0
# pytest-flask>=2.0.0
# pytest-asyncio>=0.21.0
# responses>=0.23.0
